backend/requirements.txt
modal==0.50.2632
openai==0.27.2
pydantic==1.10.7
python-dotenv==1.0.0
